requirements.txt
Flask==2.1.1
pandas==1.4.2
scikit-learn==1.0.2
google-cloud-translate==3.0.2
